# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Knodex API
  description: |
    REST API for Knodex — a Kubernetes Resource Orchestrator (KRO) visualization
    and management platform.

    Knodex provides a web interface for browsing RGD catalogs, deploying instances,
    managing projects and RBAC, and monitoring OPA Gatekeeper compliance.

    ## Authentication

    Most endpoints require a JWT Bearer token obtained via:
    - **Local login**: `POST /api/v1/auth/local/login` with username/password
    - **OIDC login**: Browser flow via `GET /api/v1/auth/oidc/login`

    Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    Tokens expire after 1 hour (HS256). There is no refresh mechanism — re-authenticate
    after expiry.

    ## Enterprise Features

    Endpoints tagged **enterprise** require an enterprise license. In OSS builds they
    return `402 Payment Required` or `404 Not Found`.

    ## Rate Limiting

    Authentication endpoints are rate-limited per IP. Protected endpoints are
    rate-limited per authenticated user (default 100 req/min).
  version: 1.0.0
  contact:
    name: Provops
    url: https://github.com/knodex/knodex
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080
    description: Local development server (Tilt)
  - url: http://knodex.knodex.svc.cluster.local:8080
    description: In-cluster service
...
paths:
...
  /api/v1/rgds/{name}/resources:
    get:
      summary: Get RGD resource graph
      description: Returns the internal resource graph (templates and external refs) within an RGD.
      tags: [rgds]
      operationId: getRGDResourceGraph
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RGDName'
        - $ref: '#/components/parameters/NamespaceQuery'
      responses:
        '200':
          description: Resource graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceGraphResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/resources:
    get:
      summary: List Kubernetes resources
      description: Lists Kubernetes resources by apiVersion, kind, and optional namespace. Used for ExternalRef selectors.
      tags: [kubernetes]
      operationId: listK8sResources
      security:
        - bearerAuth: []
      parameters:
        - name: apiVersion
          in: query
          required: true
          description: Kubernetes API version (e.g., v1, apps/v1)
          schema:
            type: string
        - name: kind
          in: query
          required: true
          description: Resource kind (e.g., ConfigMap, Secret)
          schema:
            type: string
        - $ref: '#/components/parameters/NamespaceQuery'
      responses:
        '200':
          description: Resource list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/K8sResourceListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from local login or OIDC flow (HS256, 1hr expiry)

  # -------------------------------------------------------------------------
  # Reusable parameters
  # -------------------------------------------------------------------------
  parameters:
...
    NamespaceQuery:
      name: namespace
      in: query
      description: Kubernetes namespace filter
      schema:
        type: string
    RGDName:
      name: name
      in: path
      required: true
      description: RGD resource name
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - NOT_FOUND
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - SERVICE_UNAVAILABLE
            - INTERNAL_ERROR
            - VALIDATION_FAILED
            - RATE_LIMIT_EXCEEDED
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - DEPLOYMENT_MODE_NOT_ALLOWED
            - GITOPS_ERROR
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error context (field-level errors, etc.)

    # --- Health ---
...
    ResourceGraphResponse:
      type: object
      properties:
        rgdName:
          type: string
        rgdNamespace:
          type: string
        resources:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              apiVersion:
                type: string
              kind:
                type: string
              isTemplate:
                type: boolean
              isConditional:
                type: boolean
              conditionExpr:
                type: string
              dependsOn:
                type: array
                items:
                  type: string
              externalRef:
                type: object
                properties:
                  apiVersion:
                    type: string
                  kind:
                    type: string
                  nameExpr:
                    type: string
                  namespaceExpr:
                    type: string
                  usesSchemaSpec:
                    type: boolean
                  schemaField:
                    type: string
        edges:
          type: array
          items:
            type: object
            properties:
              from:
                type: string
              to:
                type: string
              type:
                type: string
...
    K8sResourceListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        count:
          type: integer
...
security:
  - bearerAuth: []